+++ cmake/SetupCompilerOptions.cmake
@@ -263,7 +263,7 @@ endif ()
 SET( CMAKE_CXX_STANDARD_REQUIRED ON )
 
 set(BLT_CXX_STD "" CACHE STRING "Version of C++ standard")
-set_property(CACHE BLT_CXX_STD PROPERTY STRINGS c++98 c++11 c++14 c++17)
+set_property(CACHE BLT_CXX_STD PROPERTY STRINGS c++98 c++11 c++14 c++17 c++20)
 
 if (BLT_CXX_STD)
     if( BLT_CXX_STD STREQUAL c++98 ) 
@@ -296,9 +296,22 @@ if (BLT_CXX_STD)
             FLAGS_VAR CMAKE_CXX_FLAGS
             DEFAULT " "
             PGI "--c++17")
+    elseif( BLT_CXX_STD STREQUAL c++20)
+        # Error out on what does not support C++20
+        if("${CMAKE_CXX_COMPILER_ID}" STREQUAL "XL")
+            message(FATAL_ERROR "XL does not support C++20.")
+        endif()
+        if("${CMAKE_CXX_COMPILER_ID}" STREQUAL "PGI")
+            message(FATAL_ERROR "PGI is not yet supported using C++20.")
+        endif()
+        if (ENABLE_CUDA AND (NOT DEFINED CMAKE_CUDA_COMPILE_FEATURES OR (NOT "cuda_std_20" IN_LIST CMAKE_CUDA_COMPILE_FEATURES)))
+            message(FATAL_ERROR "CMake's CUDA_STANDARD does not support C++20.")
+        endif()
+
+        set(CMAKE_CXX_STANDARD 20)
     else()
         message(FATAL_ERROR "${BLT_CXX_STD} is an invalid entry for BLT_CXX_STD. "
-                            "Valid Options are ( c++98, c++11, c++14, c++17 )")
+                            "Valid Options are ( c++98, c++11, c++14, c++17, c++20 )")
     endif()
 
     if (ENABLE_CUDA)