+++ RELEASE-NOTES.md
@@ -9,6 +9,10 @@ The project release numbers follow [Semantic Versioning](http://semver.org/spec/
 
 ## [Unreleased] - Release date yyyy-mm-dd
 
+### Added
+- Added support for C++20. Note: XL does not support C++20.
+  While PGI has C++20 support, it is currently disabled (A BLT fatal error will occur).
+
 ## [Version 0.5.0] - Release date 2022-03-07
 
 ### Added